cros_build_lib: run: drop old signal workaround (reland)
Now that we enforce new enough versions in wrapper.py, there's no need
to carry these workarounds.

This is safe to land now that the autest code referring to these funcs
have been updated to not.

We also rework the logic so that if signal.signal throws a ValueError,
we assume we're in a context where we can't use signals (e.g. threads).

BUG=None
TEST=CQ passes

Change-Id: I250e917132282d110cb280e0f9e18240ad8c9079
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3529556
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>