cros lint: allow "#!/usr/bin/env python" shebangs
The reason for blocking `python` was because `python` defaulted to
`python2`, and we wanted to keep a handle on all the projects that
were actually Python 2-specific.  Now that `python` defaults to
`python3` inside the SDK & on DUTs, we can relax this restriction:
we're OK with code that uses `python` and is Python 3 only, and
with code that supports Python 2 & 3 properly.  If code is really
Python 2-only, it's going to fail inside of the SDK and on DUTs,
so there should be a good signal that it needs fixing.

BUG=chromium:997354
TEST=CQ passes

Change-Id: I2fa2f3c1af2e544200054bf89e8873a6e4ba14ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1954626
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>