cros_build_lib: run: fix mixing capture_output w/pass thru funcs
Python 3 does not allow specifying capture_output with either stdout
or stderr.  We allowed mixing where capture_output sets a default and
the others override the respective settings.  This can be useful with
pass thru functions which set a default capture_output=True and leave
the caller to override specific ones on demand.  However, in order to
be compatible with Python 3, we need to do a little more work.

In pass thru functions, we no longer default capture_output.  Instead,
we default the individual stdout & stderr to what we want iff the caller
hasn't set capture_output.  It's a bit more boilerplate, but this should
handle the various combinations correctly like Python 3 wants.

BUG=b:187789262
TEST=`./run_tests` passes

Change-Id: I21df163c5aea091073ed7818e07b4cb94ae2e9ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3890315
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>