lint: add our own pylint-2 wrapper
Add our own pylint-2 wrapper using a few of our own wheels in cipd
which work with Python 3.  We can drop this once dockerbuild updates
for Python 3 only wheels, and we can get pylint-2 in depot_tools.

This switches --py3 linting to our pylint-2.4 from the version in
the $PATH which might be anything.

BUG=chromium:1006448
TEST=`cros lint --py3` invokes our pylint-2 version

Change-Id: I90cb485df5d226d9c9e13ae9147a13a488f7f65f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2189569
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>