cros_chrome_sdk: Create build dirs when using --no-shell.
The old interactive Simple Chrome shell conveniently creates a build
dir for the given board.

This will also create build dirs in the new shell-less flow. This might
be useful for those that are accustomed to that.

Also add a couple asserts to the unittest that covers the --no-shell
functionality.

BUG=chromium:937821
TEST=unittest
TEST=simplechrome'ed with '--boards' locally

Change-Id: I06f5fa03dbc25ba2f754644f3736c839df926c48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2204569
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>