chromeos_config: set debug to true for the release-basic builder.
The release basic builder should not upload artifacts or make remote
changes. Therefore it seems like a good candidate for the debug flag.

However, the use of the debug flag is ambiguous. This CL is a test to
see how using the debug flag influences the build.

BUG=chromium:1087396
TEST=./config/refresh_generated_files ./run_tests

Change-Id: I1dff02d1c1fa2941305bd143dccf97be20f6e702
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2219699
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Navil Perez <navil@google.com>
Tested-by: Navil Perez <navil@google.com>