parser.package_info: New CPV parser module.
Transition the CPV parsing functionality to a new module.
The current location in portage_util makes it a much heavier
import than is necessary in most usages. This CL transitions
the portage_util calls themselves to just pass through to the
functions in the new module, but will be followed by a refactor
of their usages to use the new module directly, and eventually
transitioning everything to the new parse function.

BUG=chromium:1077379
TEST=run_pytest

Cq-Depend: chromium:2241587
Change-Id: Icf4fc6015c1501a403bc4c1a135e6744b91659dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2392048
Tested-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>
Reviewed-by: Michael Mortensen <mmortensen@google.com>