afdo: Post-process Android/Linux profiles to fixed number of functions
This patch adds post-process Android/Linux profiles, to prevent
the number of prod profiles grow because of any future changes. The
number of functions is selected to have 70k functions, to keep
in the same range with recent 3 merged profiles.

This change only applies to the master builder.

BUG=chromium:1062014
TEST=exp-* builders generated profiles

Change-Id: Icbf58b9849f1c05c3537a37cece3b1c7d07ed7ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2240508
Tested-by: Tiancong Wang <tcwang@google.com>
Commit-Queue: Tiancong Wang <tcwang@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>