afdo: Prepare/Bundle Android/Linux profiles in a separate handler
This is part of the efforts to formalize the usage of GS buckets.
By separating the Android/Linux profiles as a new type with a new
handler, now we can upload them to a different bucket than the
other unverified benchmark profiles.

Since the core function needs to also merge the new benchmark
profiles generated in this build, the new bundle function needs to
happen after the bundle of unverified benchmark profile.

BUG=chromium:1092541
TEST=Tested by staging-* builder

Change-Id: I6fe40208cd0ed883bf4caa36647c08ab327ce555
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2243316
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>