cros_extract_deps: Add flag to disable solver backtracking
Portage will fall back to lower versions of packages if it cannot solve
the depgraph for the highest available version of that package. We'd
prefer that be an error so that breaking a depgraph fails in the CQ
instead of silently falling back to using other versions of packages.

BUG=chromium:1094099, chromium:1067729
TEST=None

Change-Id: I5e93d21619679f62fd54aff6feb445579ac01337
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2246714
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>