cros_update: Merge into auto_updater
We don't want to have cros_update anymore as it is just a wrapper around
auto_updater. So we should move all the functionality it does into
auto_updater and make this just a wrapper for now. But since autotests
are still calling into this function, we can't yet remove cros_update
until the branch that this CL lands move entirely out of the stable
branch.

BUG=chromium:1071483
TEST=run_pytest
TEST=test_that --args="target_release=13280.0.0 target_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' source_release=13280.0.0 source_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' update_type=full" chromeos6-row4-rack9-host3.cros.corp.google.com autoupdate_EndToEndTest

Change-Id: Ia786704eb6903eafa7f75e9133b37c85cbc368bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2257395
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>