trigger_cr50_signing: Validate presence of artifact before submitting
This commit by default will reach out to ensure that the specified
artifact actually exists on the path given. It doesn't check the
content of the artifact. Also if for any reason the user can't
access the artifact an option to disable the verification check is
given.

BUG=chromium:1096113
TEST=./trigger_cr50_signing with args

Change-Id: Ia0b176b66e1a67112babf9031b700b26732f4833
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2274017
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>
Auto-Submit: George Engelbrecht <engeg@google.com>