OWNERS: add OWNERS.testplatform file
This will serve as the single OWNERS file for the testplatform team, to
be included everywhere else on the chromium Gerrit host.

I put this in chromite for consistency with the other team OWNERS files
that already live in here.

Change-Id: Ib41bc3cff4f361edae1ab99287e8e86777383051
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2288587
Tested-by: Sean Abraham <seanabraham@chromium.org>
Auto-Submit: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Commit-Queue: Dhanya Ganesh <dhanyaganesh@chromium.org>