cros_sdk: merge user PATH with root's PATH
While we want to keep entries from the user's PATH when running under
sudo, we don't want to ignore the entries that sudo/root implies.  By
default, users do not have /sbin in their PATH, but sudo/root does,
and that's where we find tools that we rely on for image management.
So instead of clobbering root's PATH, merge it instead.

BUG=chromium:1104438
TEST=`PATH=/bin:/usr/bin ./bin/cros_sdk` works again

Change-Id: I39834831e27ca75a0f2ca5b0701beacee7e56f93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2296278
Reviewed-by: Lepton Wu <lepton@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>