toolchain: Use temp bucket for unvetted artifacts on branches
Since master builders are using recipes and have migrated to use
temp bucket to store unvetted AFDO artifacts, we want to keep the
same as the branch goes to release branch. Since branch builders
are still using legacy code, we need to make this change so when
R86 is branches, the behavior will be the same.

BUG=chromium:1092541
TEST=unittest pass

Change-Id: I8ff2c4e5b920f7137db189f5fc84fea471ce0a98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2308071
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>