packages: Allow for stable and unstable ebuilds in VM uprev handlers
Currently the uprev handler only allows for a single, stable
ebuild. This makes it difficult to modify the ebuild without creating
merge conflicts with the pupr CLs. This CL changes the handler to
stabilize an unstable 9999 ebuild instead of just renaming the stable
ebuild, so changes can be landed in the unstable ebuild and pulled
into the stable ebuild by pupr.

BUG=chromium:953544
TEST=manually tested

Cq-Depend: chrome-internal:3202670
Change-Id: If7b7f2386eded8c0b74b9161843085afbbbf957c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2340622
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>