cache: Pass quiet=True to cros_build_lib.run() when extracting tarballs.
Some of the tarballs that Simple Chrome extracts have some trailing bits
after the end of the tar file. Consequently, fetching the SDK can cause
lines line "bzip2: (stdin): trailing garbage after EOF ignored" to be
printed.

These lines are mostly harmless since the contents of the tarball are
successfully extracted, and the trailing bits have no effect on
things.

So this silences all stdout/stderr from the tar command if the command
is successful. If the cmd fails, its output will be surfaced up as
part of the RunCommandError that gets thrown.

BUG=chromium:1034063
TEST=unittest
TEST=cros chrome-sdk'ed locally

Change-Id: I5e5de94b64a63c06047f4ed74272f45cff9245e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2360449
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>