cros lint: disable old-division with --py3
There's a bug in pylint when using pathlib.Path and the / operator
where it thinks we're doing division.  We can avoid this by disabling
the check when in --py3 mode.  This is safe because the import for
future division is a no-op in Python 3.

BUG=None
TEST=`cros lint --py3` on a file using pathlib no longer warns

Change-Id: I90adba5d6535ea1c39f31c1acd01e90558839252
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2377466
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>