cros_chrome_sdk: Fix args.gn in out dirs when switching between flows.
The code currently just checks that args.gn exists, and won't touch
it if it does. That can lead to problems if the args.gn file exists,
but was created via the old flow.

By regenerating it if we don't see the import() line, we should be
able to smooth over the transition.

BUG=chromium:937821
TEST=unittest

Change-Id: Ifbbdd54cc50df4e53fa206629b0a63d89b4f73a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2380406
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>