auto_updater: Remove code for getting the hostlog files.
There are two ways to get the hostlog: one in chromite and one in
autotest. Since the only user of the chromite version is
autoupdate_EndToEndTest we have updated the test to use the autotest
version and we can now delete the chromite version.

BUG=chromium:1088196
TEST=autoupdate_EndToEndTest

Change-Id: I842107e8f910395a37211ba9407d16d4c50229c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2382906
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>