cros_flash: Always log the SDK version
When using 'cros flash' from a non chromeos repo without providing
a specific version, the SDK (Chrome OS) version may not always be
obvious. This is especially true when not in the Simple Chrome
environment.

This changes a debug log message to a notice and clarifies it.

BUG=1070566
TEST=manual

Change-Id: I9e88f714d069d6058c33a6353636503d9069c4cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2386357
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>