Upload fingerprint MCU unittests tarball as artifact
This change takes the fingerprint MCU unittest binaries in a build
and uploads them as artifacts of that build. These unittest binaries
will be used in a Tast test where they are flashed to a fingerprint
MCU device in the test tab and executed on-device. We do this for
every build so that we can catch buggy CLs during a CQ run.

BUG=b:158580909
TEST=run_pytest cbuildbot/commands_unittest.py
     run_pytest service/artifacts_unittest.py

Change-Id: Ida3cb5afe4d576570de61d37100c11c387ae65fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2405306
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Yicheng Li <yichengli@chromium.org>