cros: add `help` subcommand
The lack of a `help` subcommand for `cros` catches me out regularly, so
this CL adds one that's basically an alias for `cros [subcmd] --help`.
Unfortunately it doesn't look possible to implement it using the same
pattern as the others, as it doesn't have access to the argparse
subparsers object.

BUG=none
TEST=try it with a few subcommands (e.g. `cros help flash`)

Change-Id: I1be1019d8bde37f8babf8830230660c8aa051a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2412016
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>