lddtree: add LD_ARGV0_REL
Some binaries use `/proc/self/exe` to get a link to the
currently-executing binary. Unfortunately, when `ld.so` is invoked
directly, `/proc/self/exe` alawys points to `ld.so`.

`LD_ARGV0` can only be used to determine the current executable in
programs which haven't changed their working directory from their
starting one, so that's difficult to generally use.

To solve this, this embeds the path of the current binary _relative to
ld.so_ in an env var.

BUG=chromium:1003841, chromium:1114301
TEST=inspected scripts created by cros_setup_toolchains

Change-Id: I1ef1b40206ba683439a8aca90b6812c91a85e7eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2424046
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>