cros_sdk: Add explicit --use-image option
use_image is currently the default behavior, which can be disabled by
passing --nouse-image.  In order to switch the default, callers will
need a way to explicitly request an image-backed chroot.  This adds a
--use-image option that explicitly enables the current default use_image
setting.  If both options are passed, the last one wins.

BUG=chromium:1130049
TEST=cros_sdk with --use-image and --nouse-image

Change-Id: I32470b0de1ed1dc6ba89576ecebb5954cf413361
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2431699
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>