cros_setup_toolchains: add dev-lang/rust-bootstrap
This adds the rust-bootstrap package, which bootstraps cargo and rustc
from source using a trusted compiler. The way this is intended to be
used is that the SDK builder builds this package and we publish the
resulting binaries, which we then use to build the next version
of dev-lang/rustc.

BUG=chromium:1122720
TEST=cros_setup_toolchains --show-packages host-post-cross

Change-Id: I665d72d289099c6869d63db6bfbf77df847352bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2438799
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Commit-Queue: LaMont Jones <lamontjones@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Tested-by: LaMont Jones <lamontjones@chromium.org>