toolchain: Add workaround when parsing clang version.
As discussed in crbug.com/1132918, git-r3 failed to obtain the
origin url for the LLVM checkout, resulting in the output change
of `clang --version`. This patch adds a workaround to be able to
parse a different string while the bug is being solved.

BUG=chromium:1132918
TEST=unittest and build_api run locally

Change-Id: Iadb13c31e144d954a36e74f5b77ffe31de44b301
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2441141
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Auto-Submit: Tiancong Wang <tcwang@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>