portage_testables: Let test Packages define arbitrary variables
Allow users of the Package class to specify arbitrary keys and string
values as **kwargs that will be written out to the ebuild on disk in
the synthetic test overlay. The intended use for this is to be able to
specify CROS_WORKON_* style variables to exercise those code paths in
tests.

BUG=chromium:1139412, chromium:1071391, chromium:1078251
TEST=`run_pytest`

Change-Id: I58d10290ab8628ea22973d5a6ddad38cb8a062e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2481001
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>