paygen: Make more lenient on App ID changes
It seems like we have changed the App ID on some of the boards. In that
case the paygen will fail because we have a check to make sure the App
IDs stay the same for testing purposes. But seems we don't have those
tests yet, it would be fair to remove that logic.

BUG=b:171744729
TEST=./run_tests

Change-Id: I18aceca8795335c3650ca0a10915bbf5579244c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2504356
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Auto-Submit: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>