cros_sdk: Switch use_image off by default
Switch from a default of --use-image to --nouse-image and add an
override when an existing chroot.img is present.  This results in the
following rules:

1. New chroots will not have a chroot.img by default unless --use-image
   is passed.
2. cros_sdk --replace will replace an image-backed chroot with a plain
   chroot unless --use-image is passed.
3. Existing chroots without a chroot.img will continue to work unchanged
   even if --use-image is passed (same as today).
4. Existing chroots with a chroot.img will continue to mount the
   chroot.img as needed, even if --use-image is not passed (and even if
   --nouse-image is passed).

The only rule that leads to potential surprises is #4, since there is
now no way to create a non-image chroot alongside an image-backed
chroot.  Doing so already leaves the chroot partially broken, so the
change most likely won't affect any non-accidental uses.

BUG=chromium:1130049
TEST=cros_sdk with combinations of use_image and existing chroots

Change-Id: I9a2b918f58758da6ea7613ada72cbc4a52fb69d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2515959
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>