afdo: Allow multiple Chrome ebuild to exist when getting version
When migrating to recipes and in order to allow faster executions,
toolchain builders skips setting up chroot, which makes gentoo
commands like `equery` doesn't work. This creates the need to
implement a function to get the latest info for a package. Back
then, only one ebuild is allowed to present and otherwise will
fail. However, in release branches, there could be patch 0 ebuild
in the directory, so we need to support it.

This patch adds code to find all the ebuilds belong to a package
and find the latest version by comparing all the parts.

BUG=b:171514463
TEST=unittest added and passed

Change-Id: I581af4ebf87272a0905408f284be07adfce10a27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2515019
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>