cros_build_lib: Fix ExtractTarball for uncompressed case.
Currently, running ExtractTarball on a regular, uncompressed tarball
fails because --use-compress-program expects the passed program (cat) to
take a -d flag (which cat doesn't). In the compress case, no flag is
passed so this issue doesn't happen.

Fix ExtractTarball to only add a compress program if we are
decompressing. Also, add a unit test to catch this issue.

BUG=chromium:1144993, chromium:1141320
TEST=cros flash
TEST=./chromite/run_pytest

Change-Id: I98f5428c5bda7fa7b76173330d04e58c9ac5e30f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2519236
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Stephane Belmon <sbelmon@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>