paygen (api+lib): Add override for cache dir
When executing on a bot we don't mount external cache directories so
instead let's override to a local `.paygen_cache` dir which ends up
inside the bot's named cache and is disposed on overlayfs unmount.

The alternative would be to find a way to _not_ use the cache entirely
which would work for these builders because each request is a single
shot (i.e. no cache hits).

BUG=chromium:1122854
TEST=./run_tests.py

Change-Id: Ifc3c28a7aa25b576b5b715b6de1b1024a4b3de3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2519905
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Navil Perez <navil@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>