cros_branch: Remove cros_branch
cros_branch is deprecated and is not to be used.
This change clears it from the repo to help
clean up tech debt. Documentation for the new
branch tool is located at go/cros-branch.

Testing for this change aimed to ensure
that none of the deleted code was required
by other files.

BUG=b:172362555, chromium:1128071
TEST=cros_sdk -- /mnt/host/source/chromite/run_tests

Change-Id: I083a0affcb1b5a389532f757aae6af0be99697e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2523753
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Tested-by: Julio Hurtado <juahurta@google.com>