run_pytest: improve chdir handling
The current code always chdir's to the top of the chromite dir,
but leaves the user args alone.  When trying to pass in a script,
pytest won't be able to find it anymore.  For example:
$ cd chromite/scripts
$ ./run_pytest cros_unittest.py

This will search for cros_unittest.py in the top chromite dir.

We don't actually need to chdir like this.  The code was added with
the chroot entering logic, and that was because the cros_sdk call
used the default working dir of src/scripts/ instead of telling it
to switch back to the same dir as we started in.

BUG=chromium:1147570
TEST=running tests in topdirs & subdirs still work

Change-Id: Ibab97419a635866761d20a45a77eb317ba512cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2533475
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>