api.payload: Have GeneratePayload return paths
We should have this endpoint tell us where it placed the artifacts it
generated. We'll constrain this to the local delta path as well as the
remote path if it was configured to upload.

In the future we might consider changing lib/paygen to give us a
complete listing of generated files, as some files here still escape
(i.e. are placed on gs but not returned). These include hashes and other
control files.

BUG=chromium:1122854
TEST=./run_tests.py

Change-Id: I2ba07c222980493b67b3dd22cfa3b0e87a699ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2535713
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>