nebraska: Pass 'ignore-appid' flag to Nebraska
autoupdate_StatefulCompatibility updates between <board> and
<board>-kernelnext images where the app IDs are mismatched. We need to
pass a flag to Nebraska to ignore the app ID in order for such an update
to succeed. This change adds an optional argument to
ChromiumOSUpdater.RunUpdate to ignore the app ID, and routes that to the
remote Nebraska initialization in nebraska_wrapper.py.

BUG=b:168074829
TEST=autoupdate_StatefulCompatibility with nebraska and autotest changes
in place.

Cq-Depend: chromium:2541673
Change-Id: Ie9db7055a27118515ed2eccceae9c6057ef52ea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2543204
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>