Port FindSymbolFiles from scripts/upload_symbols.py, refactor for BuildAPI.
The original script mixes finding and copying the symbols with uploading
them. Current architecture is that the BuildAPI should find symbols and
copy them to a known location, which CI can use for uploading artifacts.

This is a building block that will be used for the BundleDebugSymbols
BuildAPI endpoint.

BUG=chromium:1031380
TEST=pytest

Change-Id: Ifb66abc42f3798cb1544cc26b081999d3fdb5373
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2542576
Commit-Queue: Michael Mortensen <mmortensen@google.com>
Tested-by: Michael Mortensen <mmortensen@google.com>
Reviewed-by: Stephane Belmon <sbelmon@google.com>