cros_sdk: Move checks for snapshot compatibility
Trying to pass a snapshot command when --nouse-image is active is an
error.  With the change to making --nouse-image the default, this check
needs to be done after the automatic chroot.img detection happens so
that people don't have to explicitly pass --use-image for snapshot
operations.

BUG=chromium:1150581, chromium:1130049
TEST=cros_sdk --snapshot-list with image-backed and plain chroots

Change-Id: I70ec521ad1a8ffdd2a73284d61338f16cfd9694d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2549413
Reviewed-by: Conor McNamara <ctmcnamara@chromium.org>
Tested-by: Conor McNamara <ctmcnamara@chromium.org>
Commit-Queue: Conor McNamara <ctmcnamara@chromium.org>