Revert the revert and fix Port FindSymbolFiles from scripts/upload_symbols.py.
This reverts the revert and fixes the post-submit test failure, which
was caused by the original version of test_GatherSymbolFiles. The
error was that the output directory was in the source directory,
which could cause a race condition failure where output files were
processed as source files.

Besides correcting the test, the new method has an assertion based on
the list of SymbolFileTuple fields rather than size so that on error
the full result mismatch is displayed, enabling more insight on
failures.

This reverts commit e879dee9b4416c638125874da3fd7279a6f6465e.

BUG=chromium:1031380
TEST=pytest

Change-Id: Ic055e2c5b13432fa28de427409f2e90b23ae5e01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2551591
Reviewed-by: Stephane Belmon <sbelmon@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Michael Mortensen <mmortensen@google.com>