Introduce guest_vm_image config
When set, the BuildImage stage will repack the base and test images
into a form suitable for running as a VM inside of Chrome OS using
build_termina_image. The ArchiveStage will upload the results as a
bzipped tar file.

This new config option is enabled for the 'termina' template, which is
applied to boards tatl, tael, and sludge.

BUG=chromium:1128741
TEST=./run_pytest

Disallow-Recycled-Builds: test-failures
Change-Id: I618fcd3cab40f0bd393ed3a452072d41115e3583
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2575489
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>