Build API: Prevent entering the chroot for mock/validate calls.
The interface for mock/validate calls should allow them to all be
executed without setting up the chroot. In practice, this currently
breaks at least the dependency service, but since nothing is relying
on it, that's fine for now.

I am planning on exploring adding builders that use the mock and validate
functionality, this will be a prereq for that.

BUG=None
TEST=run_pytest

Change-Id: I21aa38a2bbc2ff2793fc7d7ba0ab8ae0236dbcb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2589369
Tested-by: Alex Klein <saklein@chromium.org>
Auto-Submit: Alex Klein <saklein@chromium.org>
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Commit-Queue: Michael Mortensen <mmortensen@google.com>