drivefs-uprev: Return UprevVersionedPackageResult instead of None
When drivefs attempts to uprev to the same version, the |uprev_result|
is returned as None. If the return from the PUpr handler is None, the
controller throws an error. The |UprevVersionedPackageResult| should be
returned to allow the property |uprevved| to return false (unless the
|add_result| method has been called, in which case it will return true.

BUG=chromium:1118212
TEST=run_pytest
TEST=https://chromeos-swarming.appspot.com/task?id=5092ea4f05c1e910

Change-Id: I0558c94135a46ebcb38abe87c7c036e689feb55e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2596696
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>