sysroot_lib: add -O to MAKEOPTS [reland]
This is a reland of 2751fd7b489ea796b428b1ed1305e8d765b73d26

The assistant code should be fixed now, and we should have a Chrome
uprev with the fix, so should be safe to land now.

Now that we've had a new enough make for a while, turn on the -O
option.  This tells make to group output from targets together and
not intermingle everything which makes reading parallel build output
a lot easier on humans.

BUG=None
TEST=CQ passes

Change-Id: I1d59aa5210903094b43036bba5b735dd72f97b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2640138
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>