cros_sdk: drop cgroups usage
In the bad old days before pid namespaces existed in the kernel, we
had problems with cbuildbot/cros_sdk leaking processes.  However, we
were able to use cpuset cgroups to keep track of all processes that
we spawned, and then destroy any dangling ones on exit.  So we used
that in cbuildbot & cros_sdk to avoid leaks.

Now that we've had the pid namespace logic in the tree for 7 years,
we should be able to assume that people have this enabled in their
kernels (as all modern distros do).  If they're using a weird setup
that doesn't have it, then they probably don't have cgroups either,
and they can get the leaking process behavior.

This allows us to delete cgroups usage from here to simplify.  It
also avoids having to fix the cgroups code to work with the new v2
layout that newer systemd versions are using.

BUG=None
TEST=`cros_sdk` works w/cgroupsv2 & systemd host

Change-Id: I5a8136c19b556dc00c63425a0a144da3044b948f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2640878
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>