osutils.SafeUnlink: fix handling of missing files in inaccessible dirs
If we try to delete a file the doesn't exist, we want to return a bool
if possible, not throw an exception.  Even more so when using sudo=True
and in a dir that only root has access to.  But the current code would
crash in sudo=True mode if the current user couldn't read the dir that
the target file is in.

For example, deleting /etc/foo/bar if /etc/foo is root:root & 700.

BUG=b:178152720
TEST=unittests pass

Change-Id: I2c37dba727aac6420742600cfb23c9a3addacfa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2648513
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>