git_unittest: untangle patch_unittest
We were reusing a fixture that created a git repo for us, but the patch
unittests need a lot more than the single git unittest does, so inline
the very limited amount of logic we actually need.  This simplifies the
code and untangles the git->patch dependency.

BUG=chromium:997354
TEST=`./run_tests` passes

Change-Id: I2f938f59f6f2ee0b05066a73216c0b2e605c8915
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2648856
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>